added caching
Implemented caching for data processing redundancy. In practice, one
may want to save time and CPU cycles by not processing data that was
already successfully processed.

Essentially the cache disables data processing for an input iff the
input already has an output and the input hash digest is equal to the
one in the cache.

Convenient command line flags for resetting or ignoring the cache
are also added.